refactor: 관리자 장소 등록 API 도메인 위치 변경
- 클래스 위치 place 패키지에서 admin 패키지로 위치 변경
- 운영 중 콘텐츠 조회 API 요청 정보에 Swagger 스키마 추가
